+++ src/plugins/system/src/android/com/foxdebug/system/System.java
@@ -10,17 +10,19 @@ import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.pm.ShortcutInfo;
 import android.content.pm.ShortcutManager;
+import android.content.res.Configuration;
 import android.graphics.Bitmap;
 import android.graphics.Color;
 import android.graphics.ImageDecoder;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.Environment;
 import android.os.PowerManager;
 import android.provider.Settings.Global;
+import android.util.Base64;
 import android.view.View;
 import android.view.Window;
+import android.view.inputmethod.InputMethodManager;
 import android.webkit.WebView;
 import androidx.core.content.FileProvider;
 import androidx.core.content.pm.ShortcutInfoCompat;
@@ -30,8 +32,12 @@ import com.foxdebug.system.BrowserDialog;
 import java.io.File;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.nio.ByteBuffer;
+import java.nio.CharBuffer;
+import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import org.apache.cordova.CallbackContext;
 import org.apache.cordova.CordovaInterface;
 import org.apache.cordova.CordovaPlugin;
@@ -89,7 +95,13 @@ public class System extends CordovaPlugin {
       case "open-in-browser":
       case "launch-app":
       case "get-global-setting":
+      case "get-available-encodings":
+      case "decode":
+      case "encode":
         break;
+      case "get-configuration":
+        getConfiguration(callbackContext);
+        return true;
       case "set-input-type":
         setInputType(arg1);
         callbackContext.success();
@@ -196,6 +208,15 @@ public class System extends CordovaPlugin {
               case "get-global-setting":
                 getGlobalSetting(arg1, callbackContext);
                 break;
+              case "get-available-encodings":
+                getAvailableEncodings(callbackContext);
+                break;
+              case "decode":
+                decode(arg1, arg2, callbackContext);
+                break;
+              case "encode":
+                encode(arg1, arg2, callbackContext);
+                break;
               default:
                 break;
             }
@@ -204,6 +225,98 @@ public class System extends CordovaPlugin {
       );
 
     return true;
+  }
+
+  private void getConfiguration(CallbackContext callback) {
+    try {
+      JSONObject result = new JSONObject();
+      Configuration config = context.getResources().getConfiguration();
+      InputMethodManager imm = (InputMethodManager) context.getSystemService(
+        Context.INPUT_METHOD_SERVICE
+      );
+      Method method =
+        InputMethodManager.class.getMethod("getInputMethodWindowVisibleHeight");
+
+      result.put("isAcceptingText", imm.isAcceptingText());
+      result.put("keyboardHeight", method.invoke(imm));
+      result.put("locale", config.locale.toString());
+      result.put("fontScale", config.fontScale);
+      result.put("keyboard", config.keyboard);
+      result.put("keyboardHidden", config.keyboardHidden);
+      result.put("hardKeyboardHidden", config.hardKeyboardHidden);
+      result.put("navigationHidden", config.navigationHidden);
+      result.put("navigation", config.navigation);
+      result.put("orientation", config.orientation);
+      callback.success(result);
+    } catch (
+      JSONException
+      | NoSuchMethodException
+      | IllegalAccessException
+      | InvocationTargetException error
+    ) {
+      callback.error(error.toString());
+    }
+  }
+
+  private void decode(
+    String content, // base64 encoded string
+    String charSetName,
+    CallbackContext callback
+  ) {
+    try {
+      byte[] bytes = Base64.decode(content, Base64.DEFAULT);
+
+      if (Charset.isSupported(charSetName) == false) {
+        callback.error("Charset not supported: " + charSetName);
+        return;
+      }
+
+      Charset charSet = Charset.forName(charSetName);
+      CharBuffer charBuffer = charSet.decode(ByteBuffer.wrap(bytes));
+      String result = String.valueOf(charBuffer);
+      callback.success(result);
+    } catch (Exception e) {
+      callback.error(e.toString());
+    }
+  }
+
+  private void encode(
+    String content, // string to encode
+    String charSetName,
+    CallbackContext callback
+  ) {
+    try {
+      if (Charset.isSupported(charSetName) == false) {
+        callback.error("Charset not supported: " + charSetName);
+        return;
+      }
+
+      Charset charSet = Charset.forName(charSetName);
+      ByteBuffer byteBuffer = charSet.encode(content);
+      byte[] bytes = new byte[byteBuffer.remaining()];
+      byteBuffer.get(bytes);
+      callback.success(bytes);
+    } catch (Exception e) {
+      callback.error(e.toString());
+    }
+  }
+
+  private void getAvailableEncodings(CallbackContext callback) {
+    try {
+      Map<String, Charset> charsets = Charset.availableCharsets();
+      JSONObject result = new JSONObject();
+      for (Map.Entry<String, Charset> entry : charsets.entrySet()) {
+        JSONObject obj = new JSONObject();
+        Charset charset = entry.getValue();
+        obj.put("label", charset.displayName());
+        obj.put("aliases", new JSONArray(charset.aliases()));
+        obj.put("name", charset.name());
+        result.put(charset.name(), obj);
+      }
+      callback.success(result);
+    } catch (Exception e) {
+      callback.error(e.toString());
+    }
   }
 
   private void requestPermissions(JSONArray arr, CallbackContext callback) {
@@ -578,7 +691,7 @@ public class System extends CordovaPlugin {
       return;
     }
 
-    callback.error("Not suppported");
+    callback.error("Not supported");
   }
 
   private void removeShortcut(String id, CallbackContext callback) {